stable version of kyber deployment scripts
